# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FileShare User API
  version: 1.0.0
  description: API для управления пользовательскими файлами и передачами.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /groups:
    post:
      tags: [Uploads]
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupCreateRequest" }
      responses:
        "201": { $ref: "#/components/responses/Group" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    Expiry: { type: string, enum: [1h, 24h, 7d, 30d, never], default: never }
    GroupCreateRequest:
      type: object
      properties:
        expiry: { $ref: "#/components/schemas/Expiry" }
        password: { type: string, maxLength: 1024 }
        maxDownloads: { type: integer, minimum: 1, maximum: 1000000 }
...
    GroupResponse:
      type: object
      required: [token, shareUrl, expiresAt, maxDownloads, hasPassword]
      properties:
        token: { type: string }
        shareUrl: { type: string, format: uri }
        expiresAt: { type: [string, "null"], format: date-time }
        maxDownloads: { type: [integer, "null"] }
        hasPassword: { type: boolean }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Group:
      description: Created group
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GroupResponse" }
...